% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_datasets.R
\docType{data}
\name{mRNA.20160128}
\alias{ACC.mRNA.20160128}
\alias{ACC.mRNA.20160128,BLCA.mRNA.20160128,BRCA.mRNA.20160128,CESC.mRNA.20160128,CHOL.mRNA.20160128,COADREAD.mRNA.20160128,DLBC.mRNA.20160128,ESCA.mRNA.20160128,GBMLGG.mRNA.20160128,HNSC.mRNA.20160128,KICH.mRNA.20160128,KIPAN.mRNA.20160128,KIRC.mRNA.20160128,KIRP.mRNA.20160128,LAML.mRNA.20160128,LIHC.mRNA.20160128,LUAD.mRNA.20160128,LUSC.mRNA.20160128,MESO.mRNA.20160128,OV.mRNA.20160128,PAAD.mRNA.20160128,PCPG.mRNA.20160128,PRAD.mRNA.20160128,SARC.mRNA.20160128,SKCM.mRNA.20160128,STAD.mRNA.20160128,STES.mRNA.20160128,TGCT.mRNA.20160128,THCA.mRNA.20160128,THYM.mRNA.20160128,UCEC.mRNA.20160128,UCS.mRNA.20160128,UVM.mRNA.20160128}
\alias{BLCA.mRNA.20160128}
\alias{BRCA.mRNA.20160128}
\alias{CESC.mRNA.20160128}
\alias{CHOL.mRNA.20160128}
\alias{COADREAD.mRNA.20160128}
\alias{DLBC.mRNA.20160128}
\alias{ESCA.mRNA.20160128}
\alias{GBMLGG.mRNA.20160128}
\alias{HNSC.mRNA.20160128}
\alias{KICH.mRNA.20160128}
\alias{KIPAN.mRNA.20160128}
\alias{KIRC.mRNA.20160128}
\alias{KIRP.mRNA.20160128}
\alias{LAML.mRNA.20160128}
\alias{LIHC.mRNA.20160128}
\alias{LUAD.mRNA.20160128}
\alias{LUSC.mRNA.20160128}
\alias{MESO.mRNA.20160128}
\alias{OV.mRNA.20160128}
\alias{PAAD.mRNA.20160128}
\alias{PCPG.mRNA.20160128}
\alias{PRAD.mRNA.20160128}
\alias{SARC.mRNA.20160128}
\alias{SKCM.mRNA.20160128}
\alias{STAD.mRNA.20160128}
\alias{STES.mRNA.20160128}
\alias{TGCT.mRNA.20160128}
\alias{THCA.mRNA.20160128}
\alias{THYM.mRNA.20160128}
\alias{UCEC.mRNA.20160128}
\alias{UCS.mRNA.20160128}
\alias{UVM.mRNA.20160128}
\alias{mRNA.20160128}
\title{mRNA datasets from TCGA project from 2016-01-28 release date}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.mRNA.20160128

BLCA.mRNA.20160128

BRCA.mRNA.20160128

CESC.mRNA.20160128

CHOL.mRNA.20160128

COADREAD.mRNA.20160128

DLBC.mRNA.20160128

ESCA.mRNA.20160128

GBMLGG.mRNA.20160128

HNSC.mRNA.20160128

KICH.mRNA.20160128

KIPAN.mRNA.20160128

KIRC.mRNA.20160128

KIRP.mRNA.20160128

LAML.mRNA.20160128

LIHC.mRNA.20160128

LUAD.mRNA.20160128

LUSC.mRNA.20160128

MESO.mRNA.20160128

OV.mRNA.20160128

PAAD.mRNA.20160128

PCPG.mRNA.20160128

PRAD.mRNA.20160128

SARC.mRNA.20160128

SKCM.mRNA.20160128

STAD.mRNA.20160128

STES.mRNA.20160128

TGCT.mRNA.20160128

THCA.mRNA.20160128

THYM.mRNA.20160128

UCEC.mRNA.20160128

UCS.mRNA.20160128

UVM.mRNA.20160128
}
\arguments{
\item{metadata}{A logical indicating whether load data into the workspace (default, \code{FALSE}) or to only display the object's metadata (\code{TRUE}). See examples.}
}
\description{
Package provides mRNA datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the release date: \code{2016-01-28} . Visit \pkg{RTCGA} site: \url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.CNV.20160128}  can be found here: \code{browseVignettes("RTCGA")}. Link to the data format explanation is in the package DESCRIPTION file.
}
\details{
\code{browseVignettes("RTCGA")}
}
\examples{

\dontrun{
 ACC.mRNA.20160128(metadata = TRUE) 
 ACC.mRNA.20160128(metadata = FALSE) 
 ACC.mRNA.20160128 
}

}
\keyword{datasets}

